[chore] 리듬 게임 HARD MODE 기능 구현
Former-commit-id: c62525224ba48b36423babfa8590dc0672e048c5

+++ BE/src/main/java/com/sugyo/domain/game/entity/FrameCoordinates.java
@@ -21,7 +21,7 @@ public class FrameCoordinates {
     private Long id;
 
     @Column(nullable = false)
-    private Long timePassed;
+    private Double timePassed;
 
     @JdbcTypeCode(SqlTypes.JSON)
     @Column(name = "frame_data", nullable = false, columnDefinition = "json")
